Changes in ActorResource, StudiResource, ActorService, StudioService

+++ src/at/technikumwien/service/ActorService.java
@@ -1,9 +1,12 @@
 package at.technikumwien.service;
 
 import at.technikumwien.entity.Actor;
+import at.technikumwien.entity.Movie;
+import at.technikumwien.resources.ActorResource;
 
 import javax.ejb.EJBException;
 import javax.ejb.Stateless;
+import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
@@ -18,6 +21,9 @@ import java.util.List;
 @Stateless
 public class ActorService {
 
+    @Inject
+    private MovieService ms;
+
     @PersistenceContext
     private EntityManager em;
 
@@ -29,11 +35,17 @@ public class ActorService {
 
     }
 
+    //CREATE
     public void importActor(Actor actor) {
-        System.out.println("Here we are");
         em.persist(actor);
     }
 
+    //READ
+    public Actor getActor(Long actorId) {
+        return em.find(Actor.class, actorId);
+    }
+
+    //UPDATE
     //so this is going to use the em.find method, which might have been a shorter way of checking the studio before :-)
     public void updateActor(Actor actor, Long actorId) {
         Actor actorOld = em.find(Actor.class, actorId);
@@ -42,21 +54,34 @@ public class ActorService {
             actorOld.setLastName(actor.getLastName());
             actorOld.setSex(actor.getSex());
             actorOld.setBirthDate(actor.getBirthDate());
-            //TODO: not sure if this needs this or an em.edit(studio) or if this line is not necessary at all:
-            em.persist(actorOld);
+            //anscheinen braucht man hier KEIN em.persist oder Ähnliches
         } else {
             throw new WebApplicationException(Response.Status.NOT_FOUND);
         }
-
     }
 
-    //same here... :-)
+    //DELETE
     public void deleteActor(Long actorId) {
         Actor actorOld = em.find(Actor.class, actorId);
         if(actorOld != null) {
-            //TODO: I'm sure I'm not allowed to delete an actor if its ID occurs in an entry in the movie_actor table
-            // TODO: does the db take care of that or what?
-            em.remove(actorOld);
+            Boolean deleteOK = true;
+            List<Movie> movieList = ms.getAllMovies();
+
+            for(Movie movie : movieList) {
+                List<Actor> actorList = movie.getActorList();
+                for (Actor actor : actorList) {
+                    if (actor.getActorId() == actorId) {
+                        deleteOK = false;
+                        break;
+                    }
+                }
+            }
+
+            if(deleteOK) {
+                em.remove(actorOld);
+            } else {
+                throw new EJBException("CANNOT DELETE STUDIO BECAUSE OF FOREIGN KEY CONSTRAINT");
+            }
         } else {
             throw new WebApplicationException(Response.Status.NOT_FOUND);
         }
@@ -64,6 +89,10 @@ public class ActorService {
 
 
 
+
+
+    //HELPER METHODS FOR CREATING MOVIE
+    //TODO: INSTEAD OF THIS, USE EM.FIND HERE AS WELL
     public Actor getActorFromDB (Actor actor) {
         Query query = em.createNamedQuery("Actor.select", Actor.class)
                 .setParameter("firstName", actor.getFirstName())
@@ -75,7 +104,6 @@ public class ActorService {
 
         return actorFromDB;
     }
-
 
     public void checkActors(List<Actor> actorList) {
         for (Actor actor : actorList) {
